papapi/requirements.txt
aiofiles==22.1.0
aiohttp==3.8.3
aiosignal==1.3.1
anyio==3.6.2
async-timeout==4.0.2
attrs==22.1.0
black==22.10.0
charset-normalizer==2.1.1
click==8.1.3
fastapi==0.88.0
frozenlist==1.3.3
h11==0.14.0
httptools==0.5.0
idna==3.4
multidict==6.0.2
mypy-extensions==0.4.3
pathspec==0.10.2
Pillow==9.3.0
platformdirs==2.5.4
pydantic==1.10.2
python-dotenv==0.21.0
PyYAML==6.0
sniffio==1.3.0
starlette==0.22.0
typing_extensions==4.4.0
uvicorn==0.20.0
uvloop==0.17.0
watchfiles==0.18.1
websockets==10.4
yarl==1.8.1
